+++ workers/unity/Packages/io.improbable.gdk.buildsystem/Configuration/BuildTargetConfig.cs
@@ -17,7 +17,18 @@ namespace Improbable.Gdk.BuildSystem.Configuration
         /// <summary>
         /// The target to build.
         /// </summary>
-        [NonSerialized] public BuildTarget Target;
+        public BuildTarget Target
+        {
+            get => buildTarget;
+            set
+            {
+                buildTarget = value;
+                if (buildTarget == BuildTarget.iOS)
+                {
+                    scriptingImplementation = ScriptingImplementation.IL2CPP;
+                }
+            }
+        }
 
         /// <summary>
         /// Is this target deprecated?
@@ -38,7 +49,15 @@ namespace Improbable.Gdk.BuildSystem.Configuration
         /// <summary>
         /// The backend scripting implementation.
         /// </summary>
-        public ScriptingImplementation ScriptingImplementation;
+        public ScriptingImplementation ScriptingImplementation
+        {
+            get => scriptingImplementation;
+            // If build target is iOS then force the Scripting Implementation to be IL2CPP (as Mono is not supported)
+            set => scriptingImplementation = Target == BuildTarget.iOS ? ScriptingImplementation.IL2CPP : value;
+        }
+
+        private BuildTarget buildTarget;
+        private ScriptingImplementation scriptingImplementation;
 
         internal string Label
         {
@@ -68,17 +87,10 @@ namespace Improbable.Gdk.BuildSystem.Configuration
         /// <summary>
         ///     Creates a new instance of a build target and its options.
         /// </summary>
-        public BuildTargetConfig(BuildTarget target, BuildOptions options,
-            bool enabled, bool required, bool deprecated = false, ScriptingImplementation scriptingImplementation = ScriptingImplementation.Mono2x)
+        public BuildTargetConfig(BuildTarget target) : this()
         {
-            Enabled = enabled;
-            Required = required;
+            ScriptingImplementation = ScriptingImplementation.Mono2x;
             Target = target;
-            Options = options;
-            Deprecated = deprecated;
-
-            // If build target is iOS then force the Scripting Implementation to be IL2CPP (as Mono is not supported)
-            ScriptingImplementation = target == BuildTarget.iOS ? ScriptingImplementation.IL2CPP : scriptingImplementation;
         }
     }
 }