+++ CHANGELOG.md
@@ -22,7 +22,7 @@
 - Upgrade Unity Burst to 1.3.5. [#1467](https://github.com/spatialos/gdk-for-unity/pull/1467)
 - Removed outline and background around component info button in the Worker Inspector. [#1468](https://github.com/spatialos/gdk-for-unity/pull/1468)
 - Refactored the `BuildContext` class. [#1461](https://github.com/spatialos/gdk-for-unity/pull/1461)
-  - Introduced a `BuildContextFilter` struct, which is required by `GetBuildContexts`.
+  - Introduced a `BuildContextSettings` struct, which is required by `GetBuildContexts`.
   - Changed `BuildConfig` class visibility from `internal` to `public`.
   - Added more testing around `BuildContext`.